@@ -2,46 +2,47 @@
   "name": "customer-client-react",
   "version": "1.0.0",
   "description": "",
-  "main": "server.js",
+  "dependencies": {
+    "@reduxjs/toolkit": "^1.9.5",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "jquery": "^3.7.1",
+    "nanoid": "^4.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.1.2",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.1"
+  },
+  "engines": {
+    "node": ">=18.0.0"
+  },
   "scripts": {
-    "start": "node server.js",
-    "lint": "eslint src",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
   "repository": {
     "type": "git",
     "url": "https://github.com/wallymathieu/customer-client-react.git"
   },
-  "author": "",
+  "author": "Oskar Gewalli",
   "license": "MIT",
-  "devDependencies": {
-    "babel-eslint": "^6.0.4",
-    "eslint": "^2.9.0",
-    "eslint-plugin-react": "^5.0.1",
-    "babel-cli": "^6.3.17",
-    "babel-core": "^6.3.17",
-    "babel-loader": "^6.2.0",
-    "babel-preset-es2015-loose": "^6.1.3",
-    "babel-preset-react": "6.3.13",
-    "babel-preset-stage-0": "^6.3.13",
-    "node-libs-browser": "^0.5.2",
-    "raw-loader": "^0.5.1",
-    "react-hot-loader": "^3.0.0-alpha.8",
-    "redux-devtools": "^3.0.1",
-    "redux-devtools-dock-monitor": "^1.0.1",
-    "redux-devtools-log-monitor": "^1.0.2",
-    "style-loader": "^0.12.3",
-    "todomvc-app-css": "^2.0.1",
-    "webpack": "^1.9.11",
-    "webpack-dev-server": "^1.9.0"
+  "eslintConfig": {
+    "extends": "react-app"
   },
-  "dependencies": {
-    "classnames": "^2.1.2",
-    "jquery": "^2.2.3",
-    "lodash": "^4.11.2",
-    "react": "^15.0.1",
-    "react-dom": "^15.0.1",
-    "react-redux": "^4.1.0",
-    "redux": "^3.1.1"
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
