@@ -1,35 +1,32 @@
-import React, { Component, PropTypes } from 'react';
-import classnames from 'classnames';
-import TextInput from './TextInput';
-import _ from 'lodash';
+import React, { useState } from "react";
+import TextInput from "./TextInput";
+import _ from "lodash";
 
-export default class Customer extends Component {
-  static propTypes = {
-    customer: PropTypes.object.isRequired,
-    editCustomer: PropTypes.func.isRequired,
+function Customer({ customer, editCustomer }) {
+  const [dirty, setDirty] = useState(false);
+
+  const handleSaveLastName = (text) => {
+    if (customer.lastName !== text) {
+      editCustomer(
+        customer.id,
+        _.extend(_.clone(customer), { lastName: text })
+      );
+      setDirty(true);
+    }
   };
-  constructor(props, context) {
-    super(props, context);
-    this.state = {
-      dirty: false
-    };
+
+  let isDirty = "";
+  if (dirty) {
+    isDirty = <div>isDirty</div>;
   }
-  handleSaveLastName(customer, text) {
-    if (customer.lastName!==text){
-      this.props.editCustomer(customer.id, _.extend(_.clone(customer),{lastName:text}));
-      this.setState({ dirty: true });
-    }
-  }
-  render() {
-    const {customer} = this.props;
-    let isDirty = '';
-    if (this.state.dirty)
-      isDirty = (<div>isDirty</div>);
-    return (<div>
-                {isDirty}
-                <div>{customer.firstName}</div>
-                <TextInput text={customer.lastName}
-                       onSave={(text) => this.handleSaveLastName(customer, text)} />
-            </div>);
-    }
+
+  return (
+    <div>
+      {isDirty}
+      <div>{customer.firstName}</div>
+      <TextInput text={customer.lastName} onSave={handleSaveLastName} />
+    </div>
+  );
 }
+
+export default Customer;
