@@ -1,34 +1,26 @@
-import React, { Component, PropTypes } from 'react';
-import classnames from 'classnames';
+import React, { useState } from "react";
 
-export default class TextInput extends Component {
-  static propTypes = {
-    onSave: PropTypes.func.isRequired,
-    text: PropTypes.string
+function TextInput({ onSave, text }) {
+  const [inputText, setInputText] = useState(text || "");
+
+  const handleChange = (e) => {
+    setInputText(e.target.value);
   };
 
-  constructor(props, context) {
-    super(props, context);
-    this.state = {
-      text: this.props.text || ''
-    };
-  }
+  const handleBlur = (e) => {
+    onSave(e.target.value);
+  };
 
-  handleChange(e) {
-    this.setState({ text: e.target.value });
-  }
-
-  handleBlur(e) {
-    this.props.onSave(e.target.value);
-  }
-
-  render() {
-    return (
-      <input type='text'
-             autoFocus='true'
-             value={this.state.text}
-             onBlur={::this.handleBlur}
-             onChange={::this.handleChange} />
-    );
-  }
+  return (
+    <input
+      type="text"
+      autoFocus={true}
+      value={inputText}
+      onBlur={handleBlur}
+      onChange={handleChange}
+    />
+  );
 }
+
+
+export default TextInput;
