@@ -1,5 +1,13 @@
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./Root.prod');
-} else {
-  module.exports = require('./Root.dev');
-}
+import React from "react";
+import { Provider } from "react-redux";
+import App from "./App";
+
+const Root = ({ store }) => {
+  return (
+    <Provider store={store}>
+      <App />
+    </Provider>
+  );
+};
+
+export default Root;
