@@ -1,5 +1,6 @@
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./configureStore.prod');
-} else {
-  module.exports = require('./configureStore.dev');
+import { createStore } from "redux";
+import rootReducer from "../reducers";
+
+export default function configureStore(initialState) {
+  return createStore(rootReducer, initialState);
 }
