@@ -1,30 +1,19 @@
 //import 'todomvc-app-css/index.css';
 import React from 'react';
-import { render } from 'react-dom';
-import { AppContainer } from 'react-hot-loader';
-import configureStore from './store/configureStore';
+import { createRoot } from "react-dom/client";
 import Root from './containers/Root';
 import loadData from './loadData';
 import syncChanges from './syncChanges';
+import configureStore from './store/configureStore';
 const store = configureStore();
 
-render(<AppContainer
-    component={Root}
-    props={{ store }}
-  />,
-  document.getElementById('root')
+
+const root = createRoot(document.getElementById("root"));
+root.render(
+  <React.StrictMode>
+    <Root store={store} />
+  </React.StrictMode>
 );
 loadData(store);
 syncChanges(store);
 
-if (module.hot) {
-  module.hot.accept('./containers/Root', () => {
-    render(
-      <AppContainer
-        component={require('./containers/Root').default}
-        props={{ store }}
-      />,
-      document.getElementById('root')
-    );
-  });
-}
