@@ -1,16 +1,13 @@
-import $ from 'jquery';
 import { parseXmlCustomers } from './adapters/Xml';
 import * as types from './constants/ActionTypes';
 
-export default function loadData(store){
-    return $.ajax({
-        url:'/CustomerService.svc/GetAllCustomers',
+export default async function loadData(store){
+    const data=await fetch('/CustomerService.svc/GetAllCustomers',{
         dataType: 'xml',
-        success:(data)=>{
-            store.dispatch({
-                type: types.ADD_CUSTOMERS,
-                infos: parseXmlCustomers(data)
-            });
-        }
+    });
+    const xml=await data.text();
+    store.dispatch({
+        type: types.ADD_CUSTOMERS,
+        infos: parseXmlCustomers(xml)
     });
 }
